gh-10: plug new filtering style modules into lib
