+++ lib/helpers/determine-filter-style.js
@@ -16,6 +16,4 @@ const determineFilterStyle = (querystring) => {
     }
 }
 
-module.exports = {
-    determineFilterStyle
-}
+module.exports = determineFilterStyle
