+++ lib/parse-filter.js
@@ -1,60 +1,32 @@
-const qs = require('qs')
+const determineFilterStyle = require('./helpers/determine-filter-style')
+const FilterStyle = require('./enums/filter-style')
+const { parseMongoFilter } = require('./mongo-filter/parse-mongo-filter')
+const { parseIbmFilter } = require('./ibm-filter/parse-ibm-filter')
 
 const parseFilter = (queryString) => {
-  const { filter } = qs.parse(queryString)
-  const comparisonObjects = []
+  let results
+  const errors = []
+  let style
 
-  if (filter) {
-    for (const [key, value] of Object.entries(filter)) {
-      const isNum = !isNaN(Number(value))
-      const isDate = !isNaN(Date.parse(value))
-      const isEqual = isNum || isDate
-      let comparator = 'ilike' // default
-      let sqlValue = value
+  try {
+    style = determineFilterStyle(queryString)
+  } catch (e) {
+    errors.push(e)
+  }
 
-      // check for explicit comparator
-      // TODO improve to catch typos/unsupported comparators
-      if (typeof value === 'object') {
-        comparator = Object.keys(value)[0] // should only be one key?
-        sqlValue = value[comparator]
-      }
-
-      // comparator conversion
-      comparator = {
-        // mongodb comparators -> objection comparators
-        $eq: '=',
-        $ne: '!=',
-        $gt: '>',
-        $gte: '>=',
-        $lt: '<',
-        $lte: '<=',
-        // TODO: add when array values are supported
-        // $in: '',
-        // $nin: '',
-        // objection comparators -> objection comparators
-        ilike: 'ilike'
-      }[comparator] || comparator
-
-      // default to "=" for dates and numbers (instead of ilike)
-      if (isEqual && comparator === 'ilike') {
-        comparator = '='
-      }
-
-      // surround value with wildcards in the case of ilike
-      if (comparator === 'ilike') {
-        sqlValue = `%${sqlValue}%`
-      }
-
-      comparisonObjects.push({ field: key, comparator, value: sqlValue })
+  if (style) {
+    const parserMap = {
+      [FilterStyle.MONGO_DB]: parseMongoFilter,
+      [FilterStyle.IBM]: parseIbmFilter,
     }
+    const styleParserReturn = parserMap[style](queryString)
+    results = styleParserReturn.results
+    errors.push(...styleParserReturn.errors)
   }
 
-  return {
-    results: comparisonObjects,
-    errors: []
-  }
+  return { results, errors }
 }
 
 module.exports = {
-  parseFilter
+  parseFilter,
 }
